Fixed shader loader refactoring cleanup
