+++ src/resources/connector_candidate.rs
@@ -1,5 +1,5 @@
-use crate::Commands;
-use bevy::prelude::Entity;
+use bevy::prelude::*;
+use bevy_prototype_lyon::entity::ShapeBundle;
 use shady_generator::Connection;
 
 #[derive(Debug)]
@@ -10,6 +10,19 @@ pub struct NodeConnectorCandidate {
 }
 
 impl NodeConnectorCandidate {
+    pub fn spawn(commands: &mut Commands, output_from: Entity, connection: Connection) -> Self {
+        Self {
+            line_entity: commands
+                .spawn_bundle(ShapeBundle {
+                    transform: Transform::from_xyz(0., 0., 1.),
+                    ..Default::default()
+                })
+                .id(),
+            output_from,
+            connection,
+        }
+    }
+
     pub fn remove_candidate(commands: &mut Commands, res: Option<&Self>) {
         if let Some(c) = res {
             commands.entity(c.line_entity).despawn();
