+++ src/systems/ui/mod.rs
@@ -1,19 +1,17 @@
 mod constants;
 pub mod creation_menu;
+pub mod log;
 mod preview_material;
 
-use crate::components::{LogElement, LogLevel};
-use crate::resources::{Candidate, IOState, OperationSelection, RenderPhase, TypeSelection};
+use crate::resources::{Candidate, IOState, OperationSelection, TypeSelection};
 use crate::systems::ui::constants::handle_constants;
 use crate::systems::ui::preview_material::handle_preview;
 use crate::{CurrentShader, PreviewMaterial, UiState, VERSION};
 use bevy::prelude::*;
-use bevy_egui::egui::{Color32, ComboBox, Frame, RichText};
+use bevy_egui::egui::{Color32, Frame, RichText};
 use bevy_egui::{egui, EguiContext};
 use shady_generator::node_operation::{NativeFunction, NativeOperation, NonScalarSwizzle};
-use shady_generator::{
-    ConstantValue, FloatingNativeType, GraphicLibrary, NativeType, NonScalarNativeType, ShaderType,
-};
+use shady_generator::{ConstantValue, FloatingNativeType, NativeType, NonScalarNativeType};
 
 pub fn setup(egui_ctx: ResMut<EguiContext>) {
     egui_ctx.ctx().set_visuals(egui::Visuals {
@@ -45,20 +43,20 @@ pub fn menu(
             ui.label("Shader name:");
             ui.text_edit_singleline(&mut shader.name);
 
-            ComboBox::from_label("Type")
-                .selected_text(shader.shader_type.to_string())
-                .show_ui(ui, |ui| {
-                    ui.selectable_value(&mut shader.shader_type, ShaderType::Vertex, "Vertex");
-                    ui.selectable_value(&mut shader.shader_type, ShaderType::Fragment, "Fragment");
-                });
+            // ComboBox::from_label("Type")
+            //     .selected_text(shader.shader_type.to_string())
+            //     .show_ui(ui, |ui| {
+            //         ui.selectable_value(&mut shader.shader_type, ShaderType::Vertex, "Vertex");
+            //         ui.selectable_value(&mut shader.shader_type, ShaderType::Fragment, "Fragment");
+            //     });
 
-            ComboBox::from_label("Target Lib")
-                .selected_text(shader.library.to_string())
-                .show_ui(ui, |ui| {
-                    ui.selectable_value(&mut shader.library, GraphicLibrary::OpenGl, "OpenGl");
-                    ui.selectable_value(&mut shader.library, GraphicLibrary::OpenGlEs, "OpenGlEs");
-                    ui.selectable_value(&mut shader.library, GraphicLibrary::WebGPU, "WebGPU");
-                });
+            // ComboBox::from_label("Target Lib")
+            //     .selected_text(shader.library.to_string())
+            //     .show_ui(ui, |ui| {
+            //         ui.selectable_value(&mut shader.library, GraphicLibrary::OpenGl, "OpenGl");
+            //         ui.selectable_value(&mut shader.library, GraphicLibrary::OpenGlEs, "OpenGlEs");
+            //         ui.selectable_value(&mut shader.library, GraphicLibrary::WebGPU, "WebGPU");
+            //     });
 
             ui.separator();
             ui.label("Properties");
@@ -128,20 +126,20 @@ pub fn menu(
                     handle_preview(ui, &mut preview_material);
                 });
             });
-            ComboBox::from_label("Render Phase")
-                .selected_text(preview_material.render_phase.to_string())
-                .show_ui(ui, |ui| {
-                    ui.selectable_value(
-                        &mut preview_material.render_phase,
-                        RenderPhase::Opaque,
-                        RenderPhase::Opaque.to_string(),
-                    );
-                    ui.selectable_value(
-                        &mut preview_material.render_phase,
-                        RenderPhase::Transparent,
-                        RenderPhase::Transparent.to_string(),
-                    );
-                });
+            // ComboBox::from_label("Render Phase")
+            //     .selected_text(preview_material.render_phase.to_string())
+            //     .show_ui(ui, |ui| {
+            //         ui.selectable_value(
+            //             &mut preview_material.render_phase,
+            //             RenderPhase::Opaque,
+            //             RenderPhase::Opaque.to_string(),
+            //         );
+            //         ui.selectable_value(
+            //             &mut preview_material.render_phase,
+            //             RenderPhase::Transparent,
+            //             RenderPhase::Transparent.to_string(),
+            //         );
+            //     });
 
             ui.with_layout(egui::Layout::bottom_up(egui::Align::Center), |ui| {
                 ui.add(egui::Hyperlink::from_label_and_url(
@@ -175,37 +173,6 @@ pub fn menu(
                 ))
                 .small();
                 ui.label(label);
-            });
-        });
-}
-
-pub fn handle_log_elements(
-    egui_ctx: ResMut<EguiContext>,
-    mut commands: Commands,
-    mut query: Query<(Entity, &mut LogElement)>,
-    time: Res<Time>,
-) {
-    let delta_time = time.delta_seconds();
-    egui::SidePanel::right("Logger")
-        .min_width(200.)
-        .frame(Frame::none())
-        .resizable(false)
-        .show(egui_ctx.ctx(), |ui| {
-            ui.with_layout(egui::Layout::bottom_up(egui::Align::Max), |ui| {
-                for (entity, mut log) in query.iter_mut() {
-                    let mut text = RichText::new(&log.message);
-                    match log.log_level {
-                        LogLevel::Info => text = text.small().color(Color32::GREEN),
-                        LogLevel::Warn | LogLevel::Error => {
-                            text = text.strong().color(Color32::RED);
-                        }
-                    };
-                    ui.label(text);
-                    log.alive_time -= delta_time;
-                    if log.alive_time <= 0.0 {
-                        commands.entity(entity).despawn();
-                    }
-                }
             });
         });
 }
